pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-janitor"
version = "0.1.0"
description = "Email Janitor is an agentic solution that cleans up your email inbox."
authors = [{name = "Will Ahern"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.23.0",
    "litellm>=1.81.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "simplegmail>=4.1.1",
    "aiosqlite>=0.19.0",
    "streamlit>=1.45.0",
    "sqlite-utils>=3.39",
    "tenacity>=9.1.2",
]

[project.scripts]
email-janitor = "email_janitor.main:run"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.14.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src/email_janitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
